+++ example/stars/game.dart
@@ -1,0 +1,15 @@
+library stars;
+
+import 'dart:html';
+import 'package:boarding/pieces.dart';
+import 'package:boarding/util.dart';
+import 'package:boarding/boarding.dart';
+
+part 'model/stars.dart';
+part 'view/board.dart';
+
+//import 'package:stars/stars.dart';
+
+main() {
+  new Board(querySelector('#canvas')).draw();
+}